Revert strange commit to protobuf
> [libprotobuf ERROR /place/db-0/key_0/user-0/svnc2/2/temp/3262ae3cc46cab740e1e98c939390932/9C/tmpiRRks9/.mount/contrib/libs/protobuf/src/google/protobuf/message_lite.cc:238] Exceeded maximum protobuf size of 2GB: 13000000000

ref:3a6ab9040d5c543846623990284ffd745d85c6bd

+++ contrib/libs/protobuf/src/google/protobuf/io/coded_stream.cc
@@ -224,7 +224,7 @@ bool CodedInputStream::SkipFallback(int count, int original_buffer_size) {
   return input_->Skip(count);
 }
 
-bool CodedInputStream::GetDirectBufferPointer(const void** data, intptr_t* size) {
+bool CodedInputStream::GetDirectBufferPointer(const void** data, int* size) {
   if (BufferSize() == 0 && !Refresh()) return false;
 
   *data = buffer_;
@@ -679,7 +679,7 @@ bool CodedOutputStream::Skip(int count) {
   return true;
 }
 
-bool CodedOutputStream::GetDirectBufferPointer(void** data, intptr_t* size) {
+bool CodedOutputStream::GetDirectBufferPointer(void** data, int* size) {
   if (buffer_size_ == 0 && !Refresh()) return false;
 
   *data = buffer_;
